Updated BaseStats to use decimal instead of double. Added a BasePlayerLevelStatics, intended to be used for rating-to-percentage (and backwards) conversions.

+++ TreeCalc/BaseStats.cs
@@ -12,25 +12,25 @@ namespace TreeCalc {
 
         public int HasteRating { get; set; }
 
-        public double HastePercentage { get; set; }
+        public decimal HastePercentage { get; set; }
 
         public int CritRating { get; set; }
 
-        public double CritPercentage { get; set; }
+        public decimal CritPercentage { get; set; }
 
         public int MasteryRating { get; set; }
 
-        public double MasteryPercentage { get; set; }
+        public decimal MasteryPercentage { get; set; }
 
         public int VersatilityRating { get; set; }
 
-        public double VersatilityPercentage { get; set; }
+        public decimal VersatilityPercentage { get; set; }
 
         public int Stamina { get; set; }
 
         public int LeechRating { get; set; }
 
-        public double LeechPercentage { get; set; }
+        public decimal LeechPercentage { get; set; }
 
         /// <summary>
         /// Spell power or attack power.
